Update navbar with github link and theme switcher
