+++ .editorconfig
@@ -7,7 +7,7 @@ end_of_line = crlf
 indent_size = 2
 indent_style = tab
 insert_final_newline = true
-max_line_length = 80
+max_line_length = 120
 trim_trailing_whitespace = true
 
 [*.md]
