+++ src/Shared/components/index.ts
@@ -1,1 +1,1 @@
-export * from './NavBar';
+export * from './NavBar/NavBar';
